requirements.txt
flask==3.0.2
PyPDF2==3.0.1
openai==1.12.0
pinecone-client==3.0.2
python-dotenv==1.0.1
werkzeug==3.0.1
tqdm==4.66.2
twilio==8.12.0
elevenlabs==0.3.0
deepgram-sdk==2.12.0
python-multipart==0.0.9
requests==2.31.0
pydub==0.25.1
sounddevice==0.4.6
soundfile==0.12.1
numpy==1.24.3
simpleaudio==1.0.4
